Move all command-related classes into itachi.command package
