+++ src/main/java/itachi/command/Command.java
@@ -1,4 +1,10 @@
+package itachi.command;
 import java.io.IOException;
+
+import itachi.ItachiException;
+import itachi.Storage;
+import itachi.Ui;
+import itachi.task.TaskList;
 
 public abstract class Command {
     public abstract void execute(TaskList tasks, Ui ui, Storage storage) throws ItachiException, IOException;
