+++ src/main/java/itachi/command/Commands.java
@@ -1,3 +1,5 @@
+package itachi.command;
+import itachi.ItachiException;
 
 public enum Commands {
     TODO,
