+++ src/main/java/itachi/command/DeadlineCommand.java
@@ -1,4 +1,12 @@
+package itachi.command;
 import java.io.IOException;
+
+import itachi.ItachiException;
+import itachi.Storage;
+import itachi.Ui;
+import itachi.command.Command;
+import itachi.task.Deadline;
+import itachi.task.TaskList;
 
 public class DeadlineCommand extends Command {
     private String description;
