+++ src/main/java/itachi/command/UnmarkCommand.java
@@ -1,9 +1,17 @@
+package itachi.command;
 import java.io.IOException;
 
-public class MarkCommand extends Command{
+import itachi.ItachiException;
+import itachi.Storage;
+import itachi.Ui;
+import itachi.command.Command;
+import itachi.task.Task;
+import itachi.task.TaskList;
+
+public class UnmarkCommand extends Command {
     private int index;
 
-    public MarkCommand(int index) {
+    public UnmarkCommand(int index) {
         this.index = index;
     }
 
@@ -17,11 +25,13 @@ public class MarkCommand extends Command{
         }
 
         Task task = tasks.get(this.index);
-        task.markAsDone();
+        task.markAsNotDone();
         storage.save(tasks.getTasks());
         ui.showLine();
-        System.out.println("Nice! I've marked this task as done:");
+        System.out.println("Nice! I've marked this task as NOT done:");
         System.out.println(task);
         ui.showLine();
     }
 }
+
+
